src/test/kotlin/com/dbiagi/authorizer/service/processor/SupermarketProcessorTest.kt
package com.dbiagi.authorizer.service.processor

import com.dbiagi.authorizer.domain.CreditType
import com.dbiagi.authorizer.domain.ResultCode
import com.dbiagi.authorizer.fixture.TransactionRequestFixture
import com.dbiagi.authorizer.service.TransactionService
import org.junit.jupiter.api.Assertions.assertEquals
import org.junit.jupiter.api.Test
import org.mockito.Mockito.mock
import org.mockito.kotlin.verify

class SupermarketProcessorTest {
    private val transactionService: TransactionService = mock()
    private val supermarketProcessor = SupermarketProcessor(transactionService)

    @Test
    fun `given a tranction request should return approved`() {
        // given
        val request = TransactionRequestFixture.getTransaction()

        // when
        supermarketProcessor.authorize(request)

        // then
        verify(transactionService).process(request, CreditType.SUPERMARKET)
    }
}
